feat: split training builder into sub modules
